refactor(engine): decompose delete/emitter.rs into focused submodules
Split the 1466-line emitter module flagged by the LoC audit (PR #4305)
into one focused submodule per concern, keeping the existing public
surface intact via the same re-exports in delete/mod.rs.

- emitter/mod.rs: DeleteEmitter struct, drain loop, dispatch logic.
- emitter/fs.rs: DeleteFs trait, RealDeleteFs, RecordingDeleteFs,
  DeleteEvent.
- emitter/policy.rs: EmitterErrorPolicy plus the
  EMITTER_PARTIAL_EXIT_CODE / EMITTER_VANISHED_EXIT_CODE constants
  and the IOERR_GENERAL / IOERR_VANISHED_ONLY bit sentinels.
- emitter/cohort.rs: CohortDeleteRecord.
- emitter/tests/{mod,dispatch,error_policy,cohort}.rs: shared test
  helpers, scaffold dispatch matrix, error-policy and cohort-log
  tests, each well under the 650-line default cap.

No behaviour change. cargo check -p engine --all-features --tests
passes.
